refactor and fix batch actions/advanced vis helpers
move advanced vis stores into their own file
use `supportsBulkActions` store instead of `bulkActionsEnabled` util
(supportsBulkActions store now calls bulkActionsEnabled)
