+++ src/lib/components/workflow/workflows-summary-table-with-filters.svelte
@@ -15,7 +15,7 @@
   import Checkbox from '$lib/holocene/checkbox.svelte';
   import { workflowTerminateEnabled } from '$lib/utilities/workflow-terminate-enabled';
   import { workflowCancelEnabled } from '$lib/utilities/workflow-cancel-enabled';
-  import { supportsAdvancedVisibilityWithOrderBy } from '$lib/stores/bulk-actions';
+  import { supportsAdvancedVisibilityWithOrderBy } from '$lib/stores/advanced-visibility';
 
   const dispatch = createEventDispatcher<{
     terminateWorkflows: undefined;
