+++ src/lib/stores/bulk-actions.ts
@@ -1,35 +1,16 @@
 import { derived } from 'svelte/store';
-import { page } from '$app/stores';
 
-import {
-  advancedVisibilityEnabled,
-  advancedVisibilityEnabledWithOrderBy,
-} from '$lib/utilities/advanced-visibility-enabled';
 import { isVersionNewer } from '$lib/utilities/version-check';
 
-import { cluster } from './cluster';
+import { supportsAdvancedVisibility } from './advanced-visibility';
 import { temporalVersion } from './versions';
-
-const isCloud = derived(
-  [page],
-  ([$page]) => $page.data?.settings?.runtimeEnvironment?.isCloud,
-);
+import { settings } from './settings';
+import { bulkActionsEnabled } from '$lib/utilities/bulk-actions-enabled';
 
 export const supportsBulkActions = derived(
-  [temporalVersion, cluster],
-  ([$temporalVersion, $cluster]) =>
+  [temporalVersion, supportsAdvancedVisibility, settings],
+  ([$temporalVersion, $supportsAdvancedVisibility, $settings]) =>
     isVersionNewer($temporalVersion, '1.18.0') &&
-    advancedVisibilityEnabled($cluster, $temporalVersion),
-);
-
-export const supportsAdvancedVisibility = derived(
-  [cluster, temporalVersion, isCloud],
-  ([$cluster, $temporalVersion, $isCloud]) =>
-    advancedVisibilityEnabled($cluster, $temporalVersion) || $isCloud,
-);
-
-export const supportsAdvancedVisibilityWithOrderBy = derived(
-  [cluster, isCloud],
-  ([$cluster, $isCloud]) =>
-    advancedVisibilityEnabledWithOrderBy($cluster) || $isCloud,
+    $supportsAdvancedVisibility &&
+    bulkActionsEnabled($settings),
 );
